Added page for the elections in 2009
This was mostly a copy&paste job from 2007 with updating some data,
especially the year.
Note, that most of the sites need an update! Especially rules.wml should
be updated.
Also note, that timeline.ics could be generated with mkical.py, but I
don't know Automake or the server setup good enough to make that happen.