bin/mail-instructions: Replaced PERL version by a Python version
I finally came around to test the script on window. Although the Python
script is a bit ugly, it works. Just some Python-2.4 compatibility was
needed. In the future, we can assume to read the Template string
directory out of the file, i.e. w/o creating $member, etc. Also, check
whether a "$" in the token confuses the Template parser.